+++ jetty-core/jetty-io/src/main/java/module-info.java
@@ -23,5 +23,4 @@ module org.eclipse.jetty.io
     exports org.eclipse.jetty.io;
     exports org.eclipse.jetty.io.content;
     exports org.eclipse.jetty.io.ssl;
-  exports org.eclipse.jetty.io.writer;
 }
